requirements.txt
# requirements.txt
# 视频总结工具依赖包列表

# 音频处理
openai-whisper>=20231117
moviepy>=1.0.3

# 网络请求
requests>=2.31.0

# 视频下载
yt-dlp>=2023.12.30

# 其他工具
pathlib2>=2.3.7; python_version < "3.4" 
# 推荐使用uv进行依赖和虚拟环境管理
# 安装: pip install uv 